Push: Develop into Master

See merge request cse1105/2019-2020/repos/oopp-group-28/repository-template!190

+++ client/src/main/java/nl/tudelft/oopp/demo/entities/BuildingHours.java
@@ -8,20 +8,20 @@ import java.util.Objects;
  */
 public class BuildingHours {
     private int id;
-    private int day;
+    private long day;
 
     private Building building;
     private LocalTime startTime;
     private LocalTime endTime;
 
     /**
-     * Constructor for building hours
-     * @param day = the day of the week in number representation (1 to 7)
-     * @param building the building
-     * @param startTime = the starting time
-     * @param endTime = the ending time
+     * Constructor for building hours.
+     * @param day = the day of the week in number representation (1 to 7).
+     * @param building the building.
+     * @param startTime = the starting time.
+     * @param endTime = the ending time.
      */
-    public BuildingHours(int day, Building building, LocalTime startTime, LocalTime endTime) {
+    public BuildingHours(long day, Building building, LocalTime startTime, LocalTime endTime) {
         this.day = day;
         this.building = building;
         this.startTime = startTime;
@@ -32,7 +32,7 @@ public class BuildingHours {
 
     }
 
-    public void setDay(int day) {
+    public void setDay(long day) {
         this.day = day;
     }
 
@@ -52,7 +52,7 @@ public class BuildingHours {
         return id;
     }
 
-    public int getDay() {
+    public long getDay() {
         return day;
     }
