+++ server/src/main/java/nl/tudelft/oopp/demo/entities/BuildingHours.java
@@ -2,6 +2,7 @@ package nl.tudelft.oopp.demo.entities;
 
 import java.time.LocalTime;
 import java.util.Objects;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -22,7 +23,7 @@ public class BuildingHours {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private int id;
-    private int day;
+    private long day;
 
     @ManyToOne
     @OnDelete(action = OnDeleteAction.CASCADE)
@@ -33,12 +34,12 @@ public class BuildingHours {
 
     /**
      * Creates a new instance of BuildingHours.
-     * @param day = the day of the week in number representation (1 to 7)
-     * @param building = the building
-     * @param startTime = the starting time
-     * @param endTime = the ending time
+     * @param day = the day of the week in number representation (1 to 7).
+     * @param building = the building.
+     * @param startTime = the starting time.
+     * @param endTime = the ending time.
      */
-    public BuildingHours(int day, Building building, LocalTime startTime, LocalTime endTime) {
+    public BuildingHours(long day, Building building, LocalTime startTime, LocalTime endTime) {
         this.day = day;
         this.building = building;
         this.startTime = startTime;
@@ -49,7 +50,7 @@ public class BuildingHours {
 
     }
 
-    public void setDay(int day) {
+    public void setDay(long day) {
         this.day = day;
     }
 
@@ -69,7 +70,7 @@ public class BuildingHours {
         return id;
     }
 
-    public int getDay() {
+    public long getDay() {
         return day;
     }
