+++ server/src/main/java/nl/tudelft/oopp/demo/entities/RestaurantHours.java
@@ -2,6 +2,7 @@ package nl.tudelft.oopp.demo.entities;
 
 import java.time.LocalTime;
 import java.util.Objects;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -22,7 +23,7 @@ public class RestaurantHours {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private int id;
-    private int day;
+    private long day;
 
     @ManyToOne
     @OnDelete(action = OnDeleteAction.CASCADE)
@@ -33,12 +34,12 @@ public class RestaurantHours {
 
     /**
      * Creates a new instance of RestaurantHours.
-     * @param day = the day of the week in number representation (1 to 7)
-     * @param restaurant = the restaurant
-     * @param startTime = the starting time
-     * @param endTime = the ending time
+     * @param day = the day of the week in number representation (1 to 7).
+     * @param restaurant = the restaurant.
+     * @param startTime = the starting time.
+     * @param endTime = the ending time.
      */
-    public RestaurantHours(int day, Restaurant restaurant, LocalTime startTime, LocalTime endTime) {
+    public RestaurantHours(long day, Restaurant restaurant, LocalTime startTime, LocalTime endTime) {
         this.day = day;
         this.restaurant = restaurant;
         this.startTime = startTime;
@@ -49,7 +50,7 @@ public class RestaurantHours {
 
     }
 
-    public void setDay(int day) {
+    public void setDay(long day) {
         this.day = day;
     }
 
@@ -69,7 +70,7 @@ public class RestaurantHours {
         return id;
     }
 
-    public int getDay() {
+    public long getDay() {
         return day;
     }
